switch to double

+++ src/kalman_filter.cpp
@@ -40,10 +40,10 @@ void KalmanFilter::Update(const VectorXd &z) {
 
 void KalmanFilter::UpdateEKF(const VectorXd &z) {
   VectorXd hx(3);
-  float px = x_(0);
-  float py = x_(1);
-  float vx = x_(2);
-  float vy = x_(3);
+  double px = x_(0);
+  double py = x_(1);
+  double vx = x_(2);
+  double vy = x_(3);
   hx << sqrt(px * px + py * py),
       atan(py / px),
       (px * vx + py * vy) / sqrt(px * px + py * py);
